perf(wip): remove extraneous functions
